Add files via upload
The last three were done in the done the same way as the others, manually, but now have strange characters for me, so they're probably not of any use.
The German parts are done to the best of my ability but the English parts are probably more accurate. I did this so that people can more easily read the content which might be useful.